+++ packages/std/src/imports.rs
@@ -28,8 +28,8 @@ extern "C" {
     #[cfg(feature = "iterator")]
     fn db_next(iterator_id: u32) -> u32;
 
-    fn canonicalize_address(source: u32, destination: u32) -> i32;
-    fn humanize_address(source: u32, destination: u32) -> i32;
+    fn canonicalize_address(source: u32, destination: u32) -> u32;
+    fn humanize_address(source: u32, destination: u32) -> u32;
 
     /// Executes a query on the chain (import). Not to be confused with the
     /// query export, which queries the state of the contract.
@@ -157,9 +157,13 @@ impl Api for ExternalApi {
         let send_ptr = &*send as *const Region as u32;
         let canon = alloc(CANONICAL_ADDRESS_BUFFER_LENGTH);
 
-        let read = unsafe { canonicalize_address(send_ptr, canon as u32) };
-        if read < 0 {
-            return Err(StdError::generic_err("canonicalize_address returned error"));
+        let result = unsafe { canonicalize_address(send_ptr, canon as u32) };
+        if result != 0 {
+            let error = unsafe { consume_string_region_written_by_vm(result as *mut Region) };
+            return Err(StdError::generic_err(format!(
+                "canonicalize_address errored: {}",
+                error
+            )));
         }
 
         let out = unsafe { consume_region(canon) };
@@ -171,16 +175,26 @@ impl Api for ExternalApi {
         let send_ptr = &*send as *const Region as u32;
         let human = alloc(HUMAN_ADDRESS_BUFFER_LENGTH);
 
-        let read = unsafe { humanize_address(send_ptr, human as u32) };
-        if read < 0 {
-            return Err(StdError::generic_err("humanize_address returned error"));
+        let result = unsafe { humanize_address(send_ptr, human as u32) };
+        if result != 0 {
+            let error = unsafe { consume_string_region_written_by_vm(result as *mut Region) };
+            return Err(StdError::generic_err(format!(
+                "humanize_address errored: {}",
+                error
+            )));
         }
 
-        let out = unsafe { consume_region(human) };
-        // we know input was correct when created, so let's save some bytes
-        let result = unsafe { String::from_utf8_unchecked(out) };
-        Ok(HumanAddr(result))
+        let address = unsafe { consume_string_region_written_by_vm(human) };
+        Ok(address.into())
     }
+}
+
+/// Takes a pointer to a Region and reads the data into a String.
+/// This is for trusted string sources only.
+unsafe fn consume_string_region_written_by_vm(from: *mut Region) -> String {
+    let data = consume_region(from);
+    // We trust the VM/chain to return correct UTF-8, so let's save some gas
+    String::from_utf8_unchecked(data)
 }
 
 /// A stateless convenience wrapper around imports provided by the VM